ass3/requirements.txt
tweepy
tqdm
json
pyyaml
transformers
pytorch-lightning
pytorch
torchmetrics
torch
wandb
scikit-learn
